# [UR][PERF] add L0, UR, and SYCL Sin Kernel Graph benchmark