```
running graph_api_benchmark_sycl SinKernelGraph graphs:0, numKernels:5, iteration 0... complete (graph_api_benchmark_sycl SinKernelGraph graphs:0, numKernels:5: 28.533 μs).
running graph_api_benchmark_sycl SinKernelGraph graphs:1, numKernels:5, iteration 0... complete (graph_api_benchmark_sycl SinKernelGraph graphs:1, numKernels:5: 45.884 μs).
running graph_api_benchmark_sycl SinKernelGraph graphs:0, numKernels:100, iteration 0... complete (graph_api_benchmark_sycl SinKernelGraph graphs:0, numKernels:100: 353.814 μs).
running graph_api_benchmark_sycl SinKernelGraph graphs:1, numKernels:100, iteration 0... complete (graph_api_benchmark_sycl SinKernelGraph graphs:1, numKernels:100: 386.979 μs).
running graph_api_benchmark_l0 SinKernelGraph graphs:0, numKernels:5, iteration 0... complete (graph_api_benchmark_l0 SinKernelGraph graphs:0, numKernels:5: 25.278 μs).
running graph_api_benchmark_l0 SinKernelGraph graphs:1, numKernels:5, iteration 0... complete (graph_api_benchmark_l0 SinKernelGraph graphs:1, numKernels:5: 20.202 μs).
running graph_api_benchmark_l0 SinKernelGraph graphs:0, numKernels:100, iteration 0... complete (graph_api_benchmark_l0 SinKernelGraph graphs:0, numKernels:100: 235.466 μs).
running graph_api_benchmark_l0 SinKernelGraph graphs:1, numKernels:100, iteration 0... complete (graph_api_benchmark_l0 SinKernelGraph graphs:1, numKernels:100: 222.501 μs).
```